requirements.txt
asgiref==3.4.1
autopep8==1.5.7
black==22.1.0
djangorestframework==3.13.1
django-crispy-forms==1.14.0
Django==3.2.5
psycopg2==2.9.3
pycodestyle==2.7.0
python-dotenv==0.19.2
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
